chore(thp): create a file for constants
