+++ suite-common/thp/src/index.ts
@@ -1,8 +1,9 @@
-export { prepareThpReducer, initialThpState, THP_BUTTON_REQUESTS_NAMES } from './thpReducer';
+export { prepareThpReducer, initialThpState } from './thpReducer';
 export type { ThpStep, ThpState } from './thpReducer';
 export { selectIsThpInProgress, selectThpStep, selectThpCredentials } from './thpSelectors';
 export { thpActions } from './thpActions';
 export * from './thpUtils';
+export { THP_BUTTON_REQUESTS_NAMES } from './thpConstants';
 export { connectThpDeviceThunk } from './connectThpDeviceThunk';
 export { finishThpAutoconnectThunk } from './finishThpAutoconnectThunk';
 export { startThpAutoconnectThunk } from './startThpAutoconnectThunk';
