+++ suite-common/thp/src/thpConstants.ts
@@ -1,0 +1,6 @@
+export const NUMBER_OF_CONNECTIONS_TO_ASK_FOR_AUTOCONNECT = 3;
+export const THP_BUTTON_REQUESTS_NAMES = [
+    'thp_pairing_request',
+    'thp_connection_request',
+    'thp_autoconnect_credential_request',
+] as const;
